modernize sha1 code
- remove fake simd
- split the generic reference implementation from the main module
